Create README.md
Hello this Github is for primarily storing the Parts to make Mine Little Pony Skins 

Including Bald Bases, Manes and Tails and finally Eyes. They all come in White/Grey to easily recolor 

Some of the skin parts are made by me and the others are pulled from 
https://github.com/MineLittlePony/Community-Skin-Pack/tree/master

Please be sure to thank the team from the github above as well if you use this skins for anything as we wouldnt be here without them! :)